+++ packages/authentication-client/test/integration/fixture.ts
@@ -14,8 +14,7 @@ export default (app: Application) => {
     entity: 'user',
     service: 'users',
     secret: 'supersecret',
-    httpStrategies: [ 'jwt' ],
-    jwtStrategies: [ 'local', 'jwt' ],
+    authStrategies: [ 'local', 'jwt' ],
     local: {
       usernameField: 'email',
       passwordField: 'password'