+++ packages/authentication-oauth/test/strategy.test.ts
@@ -4,12 +4,35 @@ import { AuthenticationService } from '@feathersjs/authentication/lib';
 
 describe('@feathersjs/authentication-oauth/strategy', () => {
   const authService: AuthenticationService = app.service('authentication');
-  const [strategy] = authService.getStrategies('test') as TestOAuthStrategy[];
+  const [ strategy ] = authService.getStrategies('test') as TestOAuthStrategy[];
 
   it('initializes, has .entityId and configuration', () => {
     assert.ok(strategy);
     assert.strictEqual(strategy.entityId, 'id');
     assert.ok(strategy.configuration.entity);
+  });
+
+  it('getRedirect', async () => {
+    app.get('authentication').oauth.redirect = '/home';
+
+    let redirect = await strategy.getRedirect({ accessToken: 'testing' });
+    assert.equal(redirect, '/home#access_token=testing');
+
+    redirect = await strategy.getRedirect(new Error('something went wrong'));
+    assert.equal(redirect, '/home#error=something%20went%20wrong');
+
+    redirect = await strategy.getRedirect(new Error());
+    assert.equal(redirect, '/home#error=OAuth%20Authentication%20not%20successful');
+
+    app.get('authentication').oauth.redirect = '/home?';
+
+    redirect = await strategy.getRedirect({ accessToken: 'testing' });
+    assert.equal(redirect, '/home?access_token=testing');
+
+    delete app.get('authentication').oauth.redirect;
+
+    redirect = await strategy.getRedirect({ accessToken: 'testing' });
+    assert.equal(redirect, null);
   });
 
   it('getProfile', async () => {
@@ -20,18 +43,6 @@ describe('@feathersjs/authentication-oauth/strategy', () => {
   });
 
   describe('authenticate', () => {
-    it('errors when strategy is not set', async () => {
-      try {
-        await strategy.authenticate({
-          id: 'newEntity'
-        }, {});
-        assert.fail('Should never get here');
-      } catch (error) {
-        assert.equal(error.name, 'NotAuthenticated');
-        assert.equal(error.message, 'Not authenticated');
-      }
-    });
-
     it('with new user', async () => {
       const authResult = await strategy.authenticate({
         strategy: 'test',