+++ packages/authentication/src/hooks/authenticate.ts
@@ -1,9 +1,8 @@
-import { flatten, merge, omit } from 'lodash';
+import { flatten, omit, merge } from 'lodash';
 import { HookContext } from '@feathersjs/feathers';
 import { NotAuthenticated } from '@feathersjs/errors';
 import Debug from 'debug';
 import { AuthenticationService } from '../service';
-import { AUTHENTICATE } from '../core';
 
 const debug = Debug('@feathersjs/authentication/hooks/authenticate');
 
@@ -45,12 +44,12 @@ export default (originalSettings: string|AuthenticateHookSettings, ...originalSt
       throw new NotAuthenticated('The authenticate hook does not need to be used on the authentication service');
     }
 
-    if (params[AUTHENTICATE] === false) {
+    if (params.authenticated === true) {
       return context;
     }
 
     if (authentication) {
-      const authParams = omit(params, 'provider', 'authentication');
+      const authParams = omit(params, 'provider', 'authentication', 'query');
 
       debug('Authenticating with', authentication, strategies);