+++ packages/express/test/authentication.test.js
@@ -116,9 +116,9 @@ describe('@feathersjs/express/authentication', () => {
       });
     });
 
-    it('errors when there are no httpStrategies', () => {
+    it('errors when there are no authStrategies', () => {
       const { accessToken } = authResult;
-      app.get('authentication').httpStrategies = [];
+      app.get('authentication').authStrategies = [];
 
       return axios.get('/dummy/dave', {
         headers: {
@@ -127,7 +127,7 @@ describe('@feathersjs/express/authentication', () => {
       }).then(() => assert.fail('Should never get here'))
       .catch(error => {
         assert.strictEqual(error.response.data.name, 'NotAuthenticated');
-        app.get('authentication').httpStrategies = [ 'jwt' ];
+        app.get('authentication').authStrategies = [ 'jwt', 'local' ];
       });
     });
 
@@ -165,7 +165,7 @@ describe('@feathersjs/express/authentication', () => {
         const { data } = error.response;
 
         assert.strictEqual(data.name, 'NotAuthenticated');
-        assert.strictEqual(data.message, 'No valid authentication strategy available');
+        assert.strictEqual(data.message, 'Invalid authentication information (no `strategy` set)');
       });
     });